Fix ring buffer wrap bug, bump to 100 messages, remap buttons
Bug: uint8_t _msgHead underflow in _viewedMsgIndex() when head
wraps to 0 after filling the ring. Caused wrong slot to be displayed
for all messages after the 10th. Fixed by casting to int.

Ring buffer bumped from 10 to 100 messages (12.8KB, 12% RAM).

Button remap by usage frequency: Middle=Enter/Approve (most used),
Ring=Escape/Kill, Pinky=Cycle Mode/BLE Switch (least used).

+++ common/display.cpp
@@ -100,7 +100,7 @@ void Display::scrollBack() {
 }
 
 int Display::_viewedMsgIndex() const {
-    return (_msgHead - 1 - _msgViewOffset + MSG_RING_SIZE * 2) % MSG_RING_SIZE;
+    return ((int)_msgHead - 1 - _msgViewOffset + MSG_RING_SIZE * 2) % MSG_RING_SIZE;
 }
 
 int Display::_wrapLineLen(const char* p) const {
